eyeball refactoring, but mostly letting eclipse do the work. 2% pita.

+++ modules/app/src/main/java/org/locationtech/jtstest/testbuilder/LayerStylePanel.java
@@ -329,7 +329,7 @@ public class LayerStylePanel extends JPanel {
       public void stateChanged(ChangeEvent e) {
         JSlider source = (JSlider)e.getSource();
         if (! source.getValueIsAdjusting()) {
-          int alpha = (int)source.getValue();
+          int alpha = source.getValue();
           geomStyle().setLineAlpha(alpha);
           JTSTestBuilder.controller().geometryViewChanged();
           JTSTestBuilder.controller().updateLayerList();
@@ -439,7 +439,7 @@ public class LayerStylePanel extends JPanel {
       public void stateChanged(ChangeEvent e) {
         JSlider source = (JSlider)e.getSource();
         if (! source.getValueIsAdjusting()) {
-          int alpha = (int)source.getValue();
+          int alpha = source.getValue();
           geomStyle().setFillAlpha(alpha);
           JTSTestBuilder.controller().geometryViewChanged();
           JTSTestBuilder.controller().updateLayerList();
