+++ modules/core/src/test/java/org/locationtech/jts/geom/impl/CoordinateArraySequenceTest.java
@@ -192,7 +192,7 @@ public class CoordinateArraySequenceTest
   private void initProgression(CoordinateSequence seq) {
     for (int index = 0; index < seq.size(); index++) {
        for( int ordinateIndex = 0; ordinateIndex < seq.getDimension(); ordinateIndex++) {
-         seq.setOrdinate(index, ordinateIndex, (double) index);
+         seq.setOrdinate(index, ordinateIndex, index);
        }
     }
   }
